Implemented Shopping Basket Functionality in JavaScript
Added JavaScript code to handle a shopping basket on a web page. Functions for adding courses to the basket, removing courses from the basket, emptying the basket, and displaying the basket in the HTML have been implemented. An array is used to keep track of courses in the basket, and the user interface is dynamically updated.